originality-engine/requirements.txt
imagehash
Pillow
datasketch
pypdf
python-docx
numpy
sentence-transformers
scikit-learn
flask
flask-cors
